# Swagger-docs/book-management-swagger-documentation.json
{"openapi":"3.0.1","info":{"title":"Book Management System API","description":"This API exposes endpoints to manage Book stufss.","termsOfService":"","contact":{"name":"SyedAlfazim","url":"","email":"syedalfazim140@gmail.com"},"license":{"name":"AEON License","url":""},"version":"1.0"},"servers":[{"url":"http://localhost:8071","description":"Server URL in Development environment"}],"tags":[{"name":"Data controller","description":"Data Management APIS"}],"paths":{"/book/update/{id}":{"put":{"tags":["Data controller","get","Data"],"summary":"update by id","description":"update data details from DB","operationId":"updateByDescription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BooksourceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{}}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/book/add":{"post":{"tags":["post","Book","Data controller"],"summary":"Register new book details","description":"adding book details into DB","operationId":"registerBook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BooksourceDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{}}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/book/retrieve/csId/{customerId}":{"get":{"tags":["Data controller","get","Data"],"summary":"retrieving data by customer id details","description":"getting data details from DB","operationId":"retrieveByCustomerID","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"customerId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{}}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/book/all":{"get":{"tags":["Data controller","get","Data"],"summary":"retrieving all data details","description":"getting data details from DB","operationId":"retrieveAllData","parameters":[{"name":"pageNo","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"application/json":{}}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}},"/book/delete/{id}":{"delete":{"tags":["Data controller","get","Data"],"summary":"delete by id","description":"delete data details from DB","operationId":"deleteByDescription","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{}}},"404":{"description":"Not Found"},"500":{"description":"Internal Server Error"}}}}},"components":{"schemas":{"BooksourceDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"trxAmmount":{"type":"number","format":"double"},"description":{"type":"string"},"trxDate":{"type":"string"},"time":{"type":"string"},"customerDetails":{"$ref":"#/components/schemas/CustomerDetails"}}},"CustomerDetails":{"type":"object","properties":{"name":{"type":"string"},"title":{"type":"string"},"age":{"type":"integer","format":"int32"}}}}}}